Rename Stream class to StringBasedStream
