+++ tests/Psr/StringBasedStream/EofTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class EofTest extends TestCase {
@@ -10,12 +10,12 @@ final class EofTest extends TestCase {
 	/**
 	 * Tests receiving true when using eof() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::eof
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::eof
 	 *
 	 * @return void
 	 */
 	public function testEofReturnsTrue() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->assertTrue($stream->eof());
 	}
