+++ tests/Psr/StringBasedStream/GetContentsTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
 use RuntimeException;
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class GetContentsTest extends TestCase {
@@ -11,15 +11,15 @@ final class GetContentsTest extends TestCase {
 	/**
 	 * Tests receiving an exception when using getContents() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::getContents
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::getContents
 	 *
 	 * @return void
 	 */
 	public function testGetContentsThrowsRuntimeException() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->expectException(RuntimeException::class);
-		$this->expectExceptionMessage(sprintf('%s::getContents() is not implemented.', Stream::class));
+		$this->expectExceptionMessage(sprintf('%s::getContents() is not implemented.', StringBasedStream::class));
 
 		$stream->getContents();
 	}
