+++ tests/Psr/StringBasedStream/GetMetadataTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
 use InvalidArgumentException;
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 use WpOrg\Requests\Tests\TypeProviderHelper;
 
@@ -12,12 +12,12 @@ final class GetMetadataTest extends TestCase {
 	/**
 	 * Tests receiving an array when using getMetadata().
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::getMetadata
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::getMetadata
 	 *
 	 * @return void
 	 */
 	public function testGetMetadataReturnsArray() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->assertSame([], $stream->getMetadata());
 	}
@@ -25,12 +25,12 @@ final class GetMetadataTest extends TestCase {
 	/**
 	 * Tests receiving null when using getMetadata().
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::getMetadata
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::getMetadata
 	 *
 	 * @return void
 	 */
 	public function testGetMetadataWithKeyReturnsNull() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->assertNull($stream->getMetadata('key'));
 	}
@@ -47,10 +47,10 @@ final class GetMetadataTest extends TestCase {
 	 * @return void
 	 */
 	public function testGetMetadataWithoutStringThrowsInvalidArgumentException($input) {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->expectException(InvalidArgumentException::class);
-		$this->expectExceptionMessage(sprintf('%s::getMetadata(): Argument #1 ($key) must be of type string', Stream::class));
+		$this->expectExceptionMessage(sprintf('%s::getMetadata(): Argument #1 ($key) must be of type string', StringBasedStream::class));
 
 		$stream->getMetadata($input);
 	}
