+++ tests/Psr/StringBasedStream/IsSeekableTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class IsSeekableTest extends TestCase {
@@ -10,12 +10,12 @@ final class IsSeekableTest extends TestCase {
 	/**
 	 * Tests receiving false when using isSeekable() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::isSeekable
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::isSeekable
 	 *
 	 * @return void
 	 */
 	public function testIsSeekableReturnsFalse() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->assertFalse($stream->isSeekable());
 	}
