+++ tests/Psr/StringBasedStream/IsWritableTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class IsWritableTest extends TestCase {
@@ -10,12 +10,12 @@ final class IsWritableTest extends TestCase {
 	/**
 	 * Tests receiving false when using isWritable() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::isWritable
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::isWritable
 	 *
 	 * @return void
 	 */
 	public function testIsWritableReturnsFalse() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->assertFalse($stream->isWritable());
 	}
