+++ tests/Psr/StringBasedStream/RewindTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
 use RuntimeException;
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class RewindTest extends TestCase {
@@ -11,15 +11,15 @@ final class RewindTest extends TestCase {
 	/**
 	 * Tests receiving an exception when using rewind() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::rewind
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::rewind
 	 *
 	 * @return void
 	 */
 	public function testRewindThrowsRuntimeException() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->expectException(RuntimeException::class);
-		$this->expectExceptionMessage(sprintf('%s::rewind() is not implemented.', Stream::class));
+		$this->expectExceptionMessage(sprintf('%s::rewind() is not implemented.', StringBasedStream::class));
 
 		$stream->rewind();
 	}
