+++ tests/Psr/StringBasedStream/SeekTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
 use RuntimeException;
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class SeekTest extends TestCase {
@@ -11,15 +11,15 @@ final class SeekTest extends TestCase {
 	/**
 	 * Tests receiving an exception when using seek() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::seek
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::seek
 	 *
 	 * @return void
 	 */
 	public function testSeekThrowsRuntimeException() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->expectException(RuntimeException::class);
-		$this->expectExceptionMessage(sprintf('%s::seek() is not implemented.', Stream::class));
+		$this->expectExceptionMessage(sprintf('%s::seek() is not implemented.', StringBasedStream::class));
 
 		$stream->seek(0);
 	}
