+++ tests/Psr/StringBasedStream/TellTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
 use RuntimeException;
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class TellTest extends TestCase {
@@ -11,15 +11,15 @@ final class TellTest extends TestCase {
 	/**
 	 * Tests receiving an exception when using tell() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::tell
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::tell
 	 *
 	 * @return void
 	 */
 	public function testTellThrowsRuntimeException() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->expectException(RuntimeException::class);
-		$this->expectExceptionMessage(sprintf('%s::tell() is not implemented.', Stream::class));
+		$this->expectExceptionMessage(sprintf('%s::tell() is not implemented.', StringBasedStream::class));
 
 		$stream->tell();
 	}
