+++ tests/Psr/StringBasedStream/WriteTest.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace WpOrg\Requests\Tests\Psr\Stream;
+namespace WpOrg\Requests\Tests\Psr\StringBasedStream;
 
 use RuntimeException;
-use WpOrg\Requests\Psr\Stream;
+use WpOrg\Requests\Psr\StringBasedStream;
 use WpOrg\Requests\Tests\TestCase;
 
 final class WriteTest extends TestCase {
@@ -11,15 +11,15 @@ final class WriteTest extends TestCase {
 	/**
 	 * Tests receiving an exception when using write() method.
 	 *
-	 * @covers \WpOrg\Requests\Psr\Stream::write
+	 * @covers \WpOrg\Requests\Psr\StringBasedStream::write
 	 *
 	 * @return void
 	 */
 	public function testWriteThrowsRuntimeException() {
-		$stream = Stream::createFromString('');
+		$stream = StringBasedStream::createFromString('');
 
 		$this->expectException(RuntimeException::class);
-		$this->expectExceptionMessage(sprintf('%s::write() is not implemented.', Stream::class));
+		$this->expectExceptionMessage(sprintf('%s::write() is not implemented.', StringBasedStream::class));
 
 		$stream->write('');
 	}
